# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FAP Expert - API de suivi de nettoyage de filtres à particules
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interventions/{id}/medias:
    get:
      operationId: listMediasIntervention
      tags: [medias]
      summary: Médias d'une intervention
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liste des médias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Media"
    post:
      operationId: ajouterMedia
      tags: [medias]
      summary: Ajouter un média (upload base64)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AjouterMediaBody"
      responses:
        "201":
          description: Média ajouté
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
  /medias/{mediaId}:
    delete:
      operationId: supprimerMedia
      tags: [medias]
      summary: Supprimer un média
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Média supprimé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Media:
      type: object
      properties:
        id:
          type: integer
        interventionId:
          type: integer
        typeMedia:
          type: string
          enum:
            - PHOTO_VEHICULE
            - PHOTO_FAP_ENTREE
            - PHOTO_ACCESSOIRES
            - VIDEO_ENDOSCOPE_ENTREE
            - PHOTO_NETTOYAGE
            - VIDEO_ENDOSCOPE_SORTIE
            - PHOTO_FAP_SORTIE
            - SIGNATURE_CLIENT
        url:
          type: string
        nomFichier:
          type: string
        mimeType:
          type: string
        tailleFichier:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - interventionId
        - typeMedia
        - url
        - nomFichier
        - mimeType
        - tailleFichier
        - createdAt
    AjouterMediaBody:
      type: object
      properties:
        typeMedia:
          type: string
          enum:
            - PHOTO_VEHICULE
            - PHOTO_FAP_ENTREE
            - PHOTO_ACCESSOIRES
            - VIDEO_ENDOSCOPE_ENTREE
            - PHOTO_NETTOYAGE
            - VIDEO_ENDOSCOPE_SORTIE
            - PHOTO_FAP_SORTIE
            - SIGNATURE_CLIENT
        nomFichier:
          type: string
        mimeType:
          type: string
        dataBase64:
          type: string
      required:
        - typeMedia
        - nomFichier
        - mimeType
        - dataBase64
...